+++ build.washingtonpost.com/components/Layout/Components/Sidebar.js
@@ -4,21 +4,25 @@ import { styled, theme, Accordion } from "@washingtonpost/wpds-ui-kit";
 import Header from "../../Typography/Headers";
 import { useRouter } from "next/router";
 import { Change } from "~/components/Markdown/Styling";
+import {
+  customNotSmBreakpoint,
+  customSmBreakpoint,
+} from "~/components/styleHelpers";
 
 const StyledAccordionRoot = styled(Accordion.Root, {
   backgroundColor: theme.colors.gray500,
-  "@sm": {
+  [customSmBreakpoint]: {
     marginBottom: "0",
   },
 });
 
 const Panel = styled("div", {
   paddingBottom: "$100",
-  "@notSm": {
+  [customNotSmBreakpoint]: {
     position: "fixed",
     height: "100%",
   },
-  "@sm": {
+  [customSmBreakpoint]: {
     width: "100%",
   },
 });
@@ -39,12 +43,11 @@ const Container = styled("div", {
     backgroundColor: theme.colors.gray200,
     borderRadius: "$round",
   },
-
   "@md": { marginTop: "0" },
-  "@notSm": {
+  [customNotSmBreakpoint]: {
     width: "300px",
   },
-  "@sm": {
+  [customSmBreakpoint]: {
     paddingBottom: 0,
   },
 });
@@ -53,7 +56,7 @@ const AccordionHeader = styled(Accordion.Header, {
   width: "100%",
   justifySelf: "flex-start",
   marginBlock: 0,
-  "@sm": {
+  [customSmBreakpoint]: {
     cursor: "pointer",
   },
 });
@@ -311,7 +314,7 @@ export default function Sidebar({ navigation, setMobileMenu }) {
           })}
         <SideBarList>
           <ListItem
-            css={{ "@notSm": { display: "none" } }}
+            css={{ [customNotSmBreakpoint]: { display: "none" } }}
             onClick={() => setMobileMenu(false)}
             isCurrent={router.asPath.includes("release-notes") ? "active" : ""}
           >