+++ packages/kit/src/theme/index.ts
@@ -1,1 +1,2 @@
 export * from "./stitches.config";
+export * from "./useResponsiveScreenSize";