+++ packages/kit/src/theme/useResponsiveScreenSize.stories.tsx
@@ -1,24 +1,33 @@
-import React from "react";
-import { within, waitFor } from "@storybook/testing-library";
+import * as React from "react";
+import { Meta, Story } from "@storybook/react";
+import { within } from "@storybook/testing-library";
 import { expect } from "@storybook/jest";
+import { Box } from "@washingtonpost/wpds-box";
 import { useResponsiveScreenSize } from "./useResponsiveScreenSize";
 
-import type { Meta, StoryObj } from "@storybook/react";
-
-const Template = () => {
-  const screenSize = useResponsiveScreenSize();
-
-  return (
-    <div>
-      <h1>Responsive Screen Size</h1>
-      <p>Current screen size: {screenSize}</p>
-    </div>
-  );
+const allModes = {
+  sm: {
+    viewport: "small",
+  },
+  md: {
+    viewport: "medium",
+  },
+  lg: {
+    viewport: "large",
+  },
+  xl: {
+    viewport: "xlarge",
+  },
+  xxl: {
+    viewport: "xxlarge",
+  },
+  minXxl: {
+    viewport: "infinity",
+  },
 };
 
-const meta: Meta<typeof Template> = {
+export default {
   title: "useResponsiveScreenSize",
-  component: Template,
   parameters: {
     viewport: {
       viewports: {
@@ -74,148 +83,143 @@ const meta: Meta<typeof Template> = {
       defaultViewport: "responsive",
     },
   },
-};
-export default meta;
+} as Meta;
 
-const allModes = {
-  sm: {
-    viewport: "small",
-  },
-  md: {
-    viewport: "medium",
-  },
-  lg: {
-    viewport: "large",
-  },
-  xl: {
-    viewport: "xlarge",
-  },
-  xxl: {
-    viewport: "xxlarge",
-  },
-  minXxl: {
-    viewport: "infinity",
-  },
+const Template: Story<typeof Box> = () => {
+  const screenSize = useResponsiveScreenSize();
+
+  return (
+    <div>
+      <h1>Responsive Screen Size</h1>
+      <p>Current screen size: {screenSize}</p>
+    </div>
+  );
 };
 
-type Story = StoryObj<typeof meta>;
+// Function to emulate pausing between interactions
+function sleep(ms) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
 
-export const InteractionsSmall: Story = {
-  parameters: {
-    viewport: {
-      defaultViewport: "small",
-    },
-    chromatic: {
-      modes: {
-        mobile: allModes["sm"],
-      },
-    },
+export const InteractionsSmall = Template.bind({});
+
+InteractionsSmall.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+  await sleep(500);
+
+  const screenSize = canvas.getByText("Current screen size: small");
+  expect(screenSize).toBeInTheDocument();
+};
+
+InteractionsSmall.parameters = {
+  viewport: {
+    defaultViewport: "small",
   },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    await waitFor(() => {
-      const screenSize = canvas.getByText("Current screen size: small");
-      expect(screenSize).toBeInTheDocument();
-    });
+  chromatic: {
+    modes: {
+      mobile: allModes["sm"],
+    },
   },
 };
 
-export const InteractionsMedium = {
-  parameters: {
-    viewport: {
-      defaultViewport: "medium",
-    },
-    chromatic: {
-      modes: {
-        tablet: allModes["md"],
-      },
-    },
+export const InteractionsMedium = Template.bind({});
+InteractionsMedium.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+  await sleep(500);
+
+  const screenSize = canvas.getByText("Current screen size: medium");
+  expect(screenSize).toBeInTheDocument();
+};
+
+InteractionsMedium.parameters = {
+  viewport: {
+    defaultViewport: "medium",
   },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    await waitFor(() => {
-      const screenSize = canvas.getByText("Current screen size: medium");
-      expect(screenSize).toBeInTheDocument();
-    });
+  chromatic: {
+    modes: {
+      tablet: allModes["md"],
+    },
   },
 };
 
-export const InteractionsLarge: Story = {
-  parameters: {
-    viewport: {
-      defaultViewport: "large",
-    },
-    chromatic: {
-      modes: {
-        tablet: allModes["lg"],
-      },
-    },
+export const InteractionsLarge = Template.bind({});
+InteractionsLarge.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+  await sleep(500);
+
+  const screenSize = canvas.getByText("Current screen size: large");
+  expect(screenSize).toBeInTheDocument();
+};
+
+InteractionsLarge.parameters = {
+  viewport: {
+    defaultViewport: "large",
   },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    await waitFor(() => {
-      const screenSize = canvas.getByText("Current screen size: large");
-      expect(screenSize).toBeInTheDocument();
-    });
+  chromatic: {
+    modes: {
+      tablet: allModes["lg"],
+    },
   },
 };
 
-export const InteractionsXLarge: Story = {
-  parameters: {
-    viewport: {
-      defaultViewport: "xlarge",
-    },
-    chromatic: {
-      modes: {
-        desktop: allModes["xl"],
-      },
-    },
+export const InteractionsXLarge = Template.bind({});
+InteractionsXLarge.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+  await sleep(500);
+
+  const screenSize = canvas.getByText("Current screen size: xlarge");
+  expect(screenSize).toBeInTheDocument();
+};
+
+InteractionsXLarge.parameters = {
+  viewport: {
+    defaultViewport: "xlarge",
   },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    await waitFor(() => {
-      const screenSize = canvas.getByText("Current screen size: xlarge");
-      expect(screenSize).toBeInTheDocument();
-    });
+  chromatic: {
+    modes: {
+      desktop: allModes["xl"],
+    },
   },
 };
 
-export const InteractionsXXLarge: Story = {
-  parameters: {
-    viewport: {
-      defaultViewport: "xxlarge",
-    },
-    chromatic: {
-      modes: {
-        desktop: allModes["xxl"],
-      },
-    },
+export const InteractionsXXLarge = Template.bind({});
+
+InteractionsXXLarge.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+  await sleep(500);
+
+  const screenSize = canvas.getByText("Current screen size: xxlarge");
+  expect(screenSize).toBeInTheDocument();
+};
+
+InteractionsXXLarge.parameters = {
+  viewport: {
+    defaultViewport: "xxlarge",
   },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    await waitFor(() => {
-      const screenSize = canvas.getByText("Current screen size: xxlarge");
-      expect(screenSize).toBeInTheDocument();
-    });
+  chromatic: {
+    modes: {
+      desktop: allModes["xxl"],
+    },
   },
 };
 
-export const InteractionsInfinity: Story = {
-  parameters: {
-    viewport: {
-      defaultViewport: "responsive",
-    },
-    chromatic: {
-      modes: {
-        desktop: allModes["minXxl"],
-      },
-    },
+export const InteractionsInfinity = Template.bind({});
+
+InteractionsInfinity.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+  await sleep(500);
+
+  const screenSize = canvas.getByText("Current screen size: infinity");
+  expect(screenSize).toBeInTheDocument();
+};
+
+InteractionsInfinity.parameters = {
+  viewport: {
+    defaultViewport: "responsive",
   },
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    await waitFor(() => {
-      const screenSize = canvas.getByText("Current screen size: infinity");
-      expect(screenSize).toBeInTheDocument();
-    });
+  chromatic: {
+    modes: {
+      desktop: allModes["minXxl"],
+    },
   },
 };